Separate module for Variables #134
