+++ python/modules/xml_parser.py
@@ -3,59 +3,9 @@ import logging, sys
 import xml.etree.ElementTree as ET
 import modules.file_managment as FM
 import modules.task as TK
+import modules.variable as VAR
 
-logger = logging.getLogger("Environment module")
-#logger = logging.getLogger("XML parser")
-
-class Variable():
-    def __init__(self, key, value, args = {}):
-        self.key = key
-        self.value = value
-        self.args = args
-
-    def get_value(self, env, args = {}):
-        return self.value
-
-    def set_value(self, value, args = {}):
-        self.value = value
-        return
-     
-    def get_variable(self, env):
-        return self
-
-class VariableReference(Variable):
-    def __init__(self, key, value, args = {}):
-        Variable.__init__(self, key, value, args = {})
-
-    def get_ref_key(self):
-        return self.value
-
-    def get_variable(self, env):
-        # creates variable with reference value
-        ref_value = self.get_value(env)
-        var = Variable(self.key, ref_value)
-        return var
-
-    def get_value(self, env, args = {}):
-        return env[self.value].get_value(env, args)
-
-class VariableList(Variable):
-    def __init__(self, key, value, args = {}):
-        Variable.__init__(self, key, value, args = {})
-
-    def get_variable(self, env):
-        # creates variable with merged value
-        merged_value = self.get_value(env)
-        var = Variable(self.key, merged_value)
-        return var
-
-    def get_value(self, env, args = {}):
-        order = sorted(self.value.keys())
-        values_list = []
-        for key in order:
-            v_part = self.value[key].get_value(env)
-            values_list.append(v_part)
-        return "".join(values_list)
+logger = logging.getLogger("XML parser")
 
 def _parse_parted_param(param_dict, param, parted_dict, p_part):
     key = param.get('key')
@@ -66,7 +16,7 @@ def _parse_parted_param(param_dict, param, parted_dict, p_part):
 
 def _add_variable_list(settings, parted_params):
     for key, params_set in parted_params.items():
-        variable = VariableList(key, parted_params[key])
+        variable = VAR.VariableList(key, parted_params[key])
         settings[key] = variable
 
 def _create_variable(param):
@@ -74,9 +24,9 @@ def _create_variable(param):
     value = param.get('value')
     p_type = param.get('type')
     if p_type == "ref" or p_type == "reference":
-        var = VariableReference(key, value)
+        var = VAR.VariableReference(key, value)
     else:
-        var = Variable(key, value)
+        var = VAR.Variable(key, value)
     return var, key
 
 def parse_param(param_dict, param, parted_dict = {}):
